+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.1.1 - 2026-05-30
+
+- fix: correct inventory preview rotation calculation
+- fix: make CopyIPCPath copy the actual preview mesh path
+
 ## 0.1.0 - 2026-05-30
 
 - Initial release